Move utility type packages into util package
