+++ README.md
@@ -124,3 +124,4 @@
 21. 파일에 몬스터 정보를 쓰고 지우는 로직의 구현을 완료하였습니다.
 22. 저장된 파일의 정보를 바탕으로 게임을 실행하는 기능을 구현하였습니다.
 23. 양수를 저장하는 클래스를 만들었습니다.
+24. util성 클래스들을 util package로 이동시켰습니다.
