+++ src/main/java/dev/idion/monsterrace/util/io/MonsterFileWriter.java
@@ -1,4 +1,4 @@
-package dev.idion.monsterrace.io;
+package dev.idion.monsterrace.util.io;
 
 import dev.idion.monsterrace.monster.Monster;
